+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Features:
 - Implement `EmulatorStart` and `EmulatorStop` for the Firebase Storage emulator.
 - Implement `EmulatorStart` and `EmulatorStop` for the Identity Platform emulator.
 - Implement `EmulatorStart` and `EmulatorStop` for the Spanner emulator.
+- Implement the `GoogleApisService`.
 
 ## v0.1.1 (2023-05-19)